requirements.txt
matplotlib==3.5.2
numpy==1.23.5
Pillow==9.2.0
Pillow==9.5.0
streamlit==1.22.0
tensorflow==2.12.0
tensorflow_intel==2.12.0